Add ruff formatter, mixed orthonormalization, tutorials website and orthogonalization methods (#30)
* Formatting Code

* Deleted Ethane and Gradient LiH example

* Refactor Code

* Replaced black and isort with Ruff

* added setuptools

* added workflow for formatting

* removed setuptools

* formatted code

* updated tequila dependency to devel branch

* updated tequila dependency to devel branch

* changed tequila dependency to main branch

* FrayedEnds description in README

* updated README.md

* updated README.md

* changed to frayedends logo in the docs

* update CI configuration to trigger linter on pull requests only

* initialize quarto

* updated gitignore

* changed to quarto

* added NWChem example

* added github workflow for quarto deplay

* added research content

* deleted contribution content

* update .gitignore to include additional Quarto build directories

* deleted intro.md

* deleted old jupyter book deploy and added quarto deploy workflow

* added output directory

* updated first example

* updated first example

* updated logo

* minor changes

* added excited states example

* execute notebooks true

* deleted old tutorial

* added new H4_excitedstates tutorial

* updated tutorial

* updated tutorials

* added new tutorial

* updated tutorials

* updated yaml

* updated PNO_FCI tutorial

* pno tutorial

* adding data for PNO tutorial

* update Quarto Deploy workflow

* updated Python version in workflow to 3.11

* removed install Frayedends in workflow

* removed research tab

* removed research tab

* update workflow

* Ortho test Pull Request (#22)

* Added correct paths to cmakelist

* Added ortho test

* Added orbital plots

* implemented mixed orthonormalization. Grouping orbitals and using linear inside the groups and between groups cholesky

* implemented new tests for comparing new orthonormalization

* added cholesky orthonormalization support

* changes on the ortho test

* changes on potentials

* Added warning function

* Added Jupyter Notebook

* Changes Jupyter Notebook

* Changes Jupyter Notebook

* Changes Jupyter Notebook

* without opt: orbital optimizer

* determinism test

* changed ortho test

* added multi run for ortho test

* changes

* Added mixed method to Integrals

* Changed mixed method to except occupation numbers

* changes on pno and integrals

* Added pno logging and plotting

* changed pno

* changed ortho test to use incremental orbitals

* pno utils implemented

* pno using FCI ansatz

* pno using SPA ansatz

* changed jupyter notebook

* changed jupyter notebook

* changed jupyter notebook

* Multi run files

* changes on orthonormalization_tutorial.ipynb

* changes on orthonormalization_tutorial.ipynb

* changes on orthonormalization_tutorial.ipynb

* changes on orthonormalization_tutorial.ipynb

* changes on orthonormalization_tutorial.ipynb

* Push Final

* Push Final

* Push Final

---------


* adding evaluate function

* adding evaluate function

* fixing botched merge

* added evaluate

* evaluate seems to work

* adding MRA basics

* removing tmp files

* ruff

* ruff

* ruff ignores notebooks

* ruff

* ruff

* ruff

* ran ruff format

* ruff: manual corrections

* added tests for evaluate

* excluded tests from ruff

* Add additional dependencies to requirements.txt

* updated pno tutorial

* Update tutorial subtitle to reflect new focus

* ruff formatter

* removed .occupation parameter and adjusted ortho tutorial

* Improve formatting of PNO usage procedure

Formatted the procedure for using pair-natural orbitals into a list format for better readability.

* fix installation, open shell doesnt work yet idk why

* small fix

* Updated tutorials and added new preview image

* Updated README

* ruff formatter

* ruff formatter

* .

---------

+++ src/open_shell/optimization_open_shell.cpp
@@ -989,15 +989,15 @@ nb::tuple Optimization_open_shell<NDIM>::get_effective_hamiltonian() {
 
 
     // Return integrals
-    Numpy2D one_e_alpha_alpha(effective_one_e_integrals[0].ptr(), {as_dims[0], as_dims[0]}, one_e_alpha_caps);
-    Numpy2D one_e_beta_beta(effective_one_e_integrals[1].ptr(), {as_dims[1], as_dims[1]}, one_e_beta_caps);
+    Numpy2D one_e_alpha_alpha(effective_one_e_integrals[0].ptr(), {static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[0])}, one_e_alpha_caps);
+    Numpy2D one_e_beta_beta(effective_one_e_integrals[1].ptr(), {static_cast<unsigned long>(as_dims[1]), static_cast<unsigned long>(as_dims[1])}, one_e_beta_caps);
     nb::list h1_list;
     h1_list.append(one_e_alpha_alpha);
     h1_list.append(one_e_beta_beta);
 
-    Numpy4D alpha_alpha(Integral_storage.as_integrals_two_body[0].ptr(), {as_dims[0], as_dims[0], as_dims[0], as_dims[0]}, two_e_aa_caps);
-    Numpy4D beta_beta(Integral_storage.as_integrals_two_body[1].ptr(), {as_dims[1], as_dims[1], as_dims[1], as_dims[1]}, two_e_bb_caps);
-    Numpy4D alpha_beta(Integral_storage.as_integrals_two_body[2].ptr(), {as_dims[0], as_dims[1], as_dims[0], as_dims[1]}, two_e_ab_caps);
+    Numpy4D alpha_alpha(Integral_storage.as_integrals_two_body[0].ptr(), {static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[0])}, two_e_aa_caps);
+    Numpy4D beta_beta(Integral_storage.as_integrals_two_body[1].ptr(), {static_cast<unsigned long>(as_dims[1]), static_cast<unsigned long>(as_dims[1]), static_cast<unsigned long>(as_dims[1]), static_cast<unsigned long>(as_dims[1])}, two_e_bb_caps);
+    Numpy4D alpha_beta(Integral_storage.as_integrals_two_body[2].ptr(), {static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[1]), static_cast<unsigned long>(as_dims[0]), static_cast<unsigned long>(as_dims[1])}, two_e_ab_caps);
     //aa, ab, bb order external; aa, bb, ab order internal
     nb::list g2_list;
     g2_list.append(alpha_alpha);
